requirements.txt
google-adk
google-cloud-aiplatform
numpy
requests
functions-framework
